+ Reset command as module
